Interpreter: Switching to sprite and removing interpreter
The previous Pixel class was not sufficient for rendering
contents for Chip8. Instead a Sprite class is now implemented
matching the specs of a Chip8 sprite, having default values for
cell_width (a pixel in Chip8 is a cell in this interpreter) which is 10
(random number with no deeper meaning) and height defaulting to 1 if
it's 0 as a Sprites height in Chip8 is N+1.

Also the previous entity generation was removed as it's now should be
replaced by the new sprite type.