Interpreter: Configurable assert for memory location access
Chip8 specifies that partials of the memory are not accessible for
'standard' programs. ensure_non_protected_access check exactly for
that. But using the interpreter with test roms, this could lead
to crashes as some test roms try to access those locations and
don't expect the interpreter to quit. Now the check for memory
access can be turned on during compile time using the flag
USE_MEM_ASSERT.