Interpreter: Reimplementation of Sprite
The Sprite type is reimplemented, this time in a separate compilation
unit as its also a separate concern. It uses a width that's 8 times
the width of a cell, where a cell is equivalent to a pixel in the
Chip8 definition. The height is at least 1 by definition as Chip8
defines the height of a Sprite as N+1 so if N = 0 the height is equal
to 1.