Interpreter: Using real rom size for loading rom
Before the rom size (max size) was fixed at 1 << 12. To be more
resource efficient the buffer size is now calculated by opening the
file, seeking to the end, and than getting the size from end to
beginning of the file.
As we need the buffer just for a few lines and taking immediatly
care of freeing it, it should be ok to
use a raw char[] pointer here.