Interpreter[Memory]: Initialize memory properly
Before only the reset_memory function was used inside of the
constructor to set the memory locations to 0. To have clean code
the m_memory field should either be initialized directly or by using
the initializer list in the constructor (RAII). This is now done by
directly initializing the memory.